Add a context manager for safe loading of data
